% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_estimation.R
\name{betacoal.maxlik}
\alias{betacoal.maxlik}
\title{Maximum-likelihood astimation of alpha}
\usage{
betacoal.maxlik(phylo, epsilon = 0)
}
\arguments{
\item{phylo}{An input phylogeny}

\item{epsilon}{Grouping parameter, as defined originally in (Strimmer, Pybus 2001)}
}
\description{
Performs a maximum-likelihood estimation of the alpha parameter, with the multifurcating skyline
as demographic function.
}
