% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{branching.times.samp}
\alias{branching.times.samp}
\title{Branching times from a (multifurcating) Phylogenetic Tree}
\usage{
branching.times.samp(phy)
}
\arguments{
\item{phy}{A phylogeny}
}
\description{
Extension of the branching.times function in ape. Takes a phylogeny (class phylo) and returns a vector with all event times in backwards time
including the coalescences and the sampling events.
}
