% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{coalescent.intervals.multi}
\alias{coalescent.intervals.multi}
\title{Function that computes the coalescent intervals of multifurcating trees}
\usage{
coalescent.intervals.multi(x)
}
\arguments{
\item{x}{Phylogeny}
}
\description{
Extension of the coalescent.intervals function in ape. Takes a phylogeny (class phylo) as input and returns a coalescentIntervals object,
which lists the interval lengths, in backwards time, along with the number of lineages during each of those intervals. Contrary to the ape
function, this can deal with multifurcations, as well as serially-sampled (non-ultrametric) trees.
}
