% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{di2multi.cons}
\alias{di2multi.cons}
\title{Transforming a binary phylogeny into a multifurcating one}
\usage{
di2multi.cons(phylo, tol = 1e-08)
}
\arguments{
\item{phylo}{A phylogeny}

\item{tol}{The collapsing parameter}
}
\description{
Transforms a binary phylogeny into a multifurcating one, collapsing all branches shorter than tol
}
