% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LambdaSkyline.R
\docType{package}
\name{LambdaSkyline-package}
\alias{LambdaSkyline}
\alias{LambdaSkyline-package}
\title{LambdaSkyline: Demographic Analysis of Multifurcating Phylogenies}
\description{
Extends skyline functions used in ape to the case of multifurcating trees using Beta(2-alpha,alpha)-coalescents.
}
\details{
Bla.
}
