% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_estimation.R
\name{lik_landscape}
\alias{lik_landscape}
\title{Likelihood landscape of a phylogenetic tree}
\usage{
lik_landscape(phylo, epsilon = 0)
}
\arguments{
\item{phylo}{An input phylogeny}

\item{epsilon}{Grouping parameter, as defined originally in (Strimmer, Pybus 2001)}
}
\description{
Computes and plots the likelihood landscape of a given phylogeny as a function of alpha
}
