% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simcoal}
\alias{simcoal}
\title{Simulate Beta-coalescents with varying population size}
\usage{
simcoal(n, alpha, N, pop_size)
}
\arguments{
\item{n}{Number of sampled lineages at present time}

\item{alpha}{The parameter of the Beta(2-alpha,alpha)-coalescent}

\item{N}{Scaling parameter of the population size}

\item{pop_size}{Vector containing the population size factor (true population size at time i is pop_size[i]*N)}
}
\description{
This function simulates the ancestral process of a sample of n lineages in a larger population which has size of order
N and fluctuates according to pop_size. The simulation is based on the results of (Schweinsberg 2003) and relies
on supercritical branching processes with heavy-tailed offspring distribution.
}
