% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_estimation.R
\name{skyline.multi.coalescentIntervals}
\alias{skyline.multi.coalescentIntervals}
\title{Multifurcating skyline plot estimate of effective population size}
\usage{
skyline.multi.coalescentIntervals(Inter, alpha, epsilon = 0)
}
\arguments{
\item{Inter}{A coalescent interval object}

\item{alpha}{The alpha parameter of the Beta(2-alpha,alpha)-coalescent}

\item{epsilon}{Grouping parameter, as defined originally in (Strimmer, Pybus 2001)}
}
\description{
Main estimation function : takes as input a coalescentIntervals object, as well as the alpha parameter of the Beta coalescent, and the grouping parameter
epsilon, and returns a skyline object, along with its log-likelihood
}
