% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_estimation.R
\name{skyline.multi.phylo}
\alias{skyline.multi.phylo}
\title{Multifurcating skyline plot estimate of effective population size}
\usage{
skyline.multi.phylo(phylo, alpha, epsilon = 0)
}
\arguments{
\item{phylo}{An input phylogeny}

\item{alpha}{The alpha parameter of the Beta(2-alpha,alpha)-coalescent}

\item{epsilon}{Grouping parameter, as defined originally in (Strimmer, Pybus 2001)}
}
\description{
Wrapper function for phylo objects
}
