% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{total_coal_rate}
\alias{total_coal_rate}
\title{Computes total rate of coalescence}
\usage{
total_coal_rate(n, alpha)
}
\arguments{
\item{n}{Number of current lineages}

\item{alpha}{Parameter of the Beta(2-alpha,alpha)-coalescent}
}
\description{
This function computes the total rate of coalescence for n lineages under the Beta(2-alpha,alpha)-coalescent
}
